+++ flexutils/viewers/__init__.py
@@ -28,8 +28,8 @@ from .viewers_data import FlexDataViewer
 
 # from .viewer_morph_salesman import FlexMorphSalesmanViewer
 from .viewer_show_structures import FlexShowStructuresViewer
-from .viewer_show_maps import FlexShowMapsViewer
-from .viewer_optimal_clusters import FlexShowOptimalClustersViewer
+from flexutils.viewers.map_model_viewers.viewer_show_maps import FlexShowMapsViewer
+from flexutils.viewers.clustering_viewers.viewer_optimal_clusters import FlexShowOptimalClustersViewer
 
 # Flexibility Xmipp viewers (offered here due to strong package dependency)
 from .xmipp import *
