+++ flexutils/viewers/annotation_3d_tools/viewer_3d_pc.py
@@ -28,13 +28,12 @@
 import numpy as np
 import os
 import pickle
+from xmipp_metadata.image_handler import ImageHandler
 
 from sklearn.neighbors import KDTree
 from sklearn.cluster import KMeans
 
-import pyvista as pv
-
-from traits.api import HasTraits, Instance, Array, Float, Int, Bool, String,\
+from traits.api import HasTraits, Instance, Array, Float, String,\
     on_trait_change, Callable
 from traits.trait_types import Button, Enum
 from traitsui.api import View, Item, HGroup, Group, HSplit, VGroup, RangeEditor, TextEditor, EnumEditor
@@ -42,25 +41,19 @@ from traitsui.api import View, Item, HGroup, Group, HSplit, VGroup, RangeEditor,
 from pyface.image_resource import ImageResource
 
 from tvtk.api import tvtk
-from tvtk.pyface.scene import Scene
 
 from mayavi import mlab
-from mayavi.core.api import PipelineBase, Source
+from mayavi.core.api import PipelineBase
 from mayavi.core.ui.api import SceneEditor, MayaviScene, \
     MlabSceneModel
 
-from pwem.emlib.image import ImageHandler
-from pyworkflow.utils import getExt
-
-import xmipp3
-
 import flexutils
-from flexutils.viewers.viewer_morph_chimerax import FlexMorphChimeraX
+from flexutils.viewers.chimera_viewers.viewer_morph_chimerax import FlexMorphChimeraX
 
 
 ################################################################################
 # The object implementing the dialog
-class VolumeSlicer(HasTraits):
+class PointCloudView(HasTraits):
     # The data to plot
     data = Array()
     grid = Array()
@@ -70,25 +63,14 @@ class VolumeSlicer(HasTraits):
 
     # The 4 views displayed (Zernike space)
     scene3d = Instance(MlabSceneModel, ())
-    scene_x = Instance(MlabSceneModel, ())
-    scene_y = Instance(MlabSceneModel, ())
-    scene_z = Instance(MlabSceneModel, ())
 
     # Real time conformation view
     scene_c = Instance(MlabSceneModel, ())
 
-    # The data source
-    data_src3d = Instance(Source)
-
-    # The image plane widgets of the 3D scene
-    ipw_3d_x = Instance(PipelineBase)
-    ipw_3d_y = Instance(PipelineBase)
-    ipw_3d_z = Instance(PipelineBase)
-
     # Line cuts for 3d view
-    line_3d_x = Instance(PipelineBase)
-    line_3d_y = Instance(PipelineBase)
-    line_3d_z = Instance(PipelineBase)
+    # line_3d_x = Instance(PipelineBase)
+    # line_3d_y = Instance(PipelineBase)
+    # line_3d_z = Instance(PipelineBase)
 
     # Point cloud (actor)
     ipw_pc = Instance(PipelineBase)
@@ -98,7 +80,6 @@ class VolumeSlicer(HasTraits):
     ipw_map = Instance(PipelineBase)
 
     # Visualization style
-    show_cut_planes = Bool(True)
     opacity = Float(0.0)
     threshold = Float(0.0)
     op_min = Float(0.0)
@@ -109,7 +90,7 @@ class VolumeSlicer(HasTraits):
     # Output path
     path = String()
 
-    # Zernike parameters
+    # Space parameters
     z_space = Array()
     vector_file = String('z_vw.txt')
 
@@ -120,24 +101,16 @@ class VolumeSlicer(HasTraits):
     map = Array()
     generated_map = Array()
 
-    # Mask
-    mask_file = String()
-
     # KMeans
     n_clusters = String("10")
 
     # Selections
-    selection_name = String()
-    # show_selections = Bool(True)
     current_sel = String('origin')
     ipw_label = Instance(PipelineBase)
     save_file = String('saved_selections.txt')
 
     # Save selections
-    add_selection = Button("Add conformation")
     save_selections = Button("Save selections")
-    remove_selection = Button("Remove conformation")
-    rename_selection = Button("Rename selection")
     compute_kmeans = Button("Compute KMeans")
 
     # Generate map function
@@ -146,7 +119,7 @@ class VolumeSlicer(HasTraits):
 
     # Points properties
     scale_pts = Button("Scale spheres")
-    scale_factor = String("3")
+    scale_factor = String("10")
 
     # ChimeraX morphing
     morphing_choice = Enum('Salesman', 'Random walk')
@@ -156,27 +129,24 @@ class VolumeSlicer(HasTraits):
     # ---------------------------------------------------------------------------
     def __init__(self, **traits):
         self.class_inputs = traits
-        super(VolumeSlicer, self).__init__(**traits)
-        # Force the creation of the image_plane_widgets:
-        self.interpolateGrid()
-        self.moveDataToGrid()
+        super(PointCloudView, self).__init__(**traits)
         self.map
-        self.ipw_3d_x
-        self.ipw_3d_y
-        self.ipw_3d_z
         self.ipw_pc
         self.ipw_map
-        self.line_3d_x
-        self.line_3d_y
-        self.line_3d_z
+        # self.line_3d_x
+        # self.line_3d_y
+        # self.line_3d_z
         self.path
+
+        # Scale data to fix in box of side 300
+        bounding = np.amax(np.abs(np.amax(self.data, axis=0) - np.amin(self.data, axis=0)))
+        self.data = (300 / bounding) * self.data
 
         # Create KDTree
         self.kdtree_data = KDTree(self.data)
-        self.kdtree_z_space = KDTree(self.z_space)
+        self.kdtree_z_pace = KDTree(self.z_space)
 
         # Generate map
-        self.mode
         if self.mode == "Zernike3D":
             from flexutils.protocols.xmipp.utils.utils import applyDeformationField
             self.generate_map = applyDeformationField
@@ -209,58 +179,38 @@ class VolumeSlicer(HasTraits):
     # ---------------------------------------------------------------------------
     # Default values
     # ---------------------------------------------------------------------------
-    def _data_src3d_default(self):
-        return mlab.pipeline.scalar_field(self.grid, figure=self.scene3d.mayavi_scene, colormap="viridis")
+    # def make_line_3d(self, axis_name):
+    #     if axis_name == "x":
+    #         normal = np.array([1, 0, 0])
+    #     elif axis_name == "y":
+    #         normal = np.array([0, 1, 0])
+    #     elif axis_name == "z":
+    #         normal = np.array([0, 0, 1])
+    #     white = (1, 1, 1)
+    #     grid_size = np.asarray(self.grid.shape)
+    #     point1 = 0.5 * grid_size * (1 - normal)
+    #     point2 = grid_size * normal + point1
+    #     x = [point1[0], point2[0]]
+    #     y = [point1[1], point2[1]]
+    #     z = [point1[2], point2[2]]
+    #     line = mlab.plot3d(x, y, z, color=white, tube_radius=1., figure=self.scene3d.mayavi_scene)
+    #     return line
 
-    def make_ipw_3d(self, axis_name):
-        ipw = mlab.pipeline.image_plane_widget(self.data_src3d,
-                                               figure=self.scene3d.mayavi_scene,
-                                               transparent=True,
-                                               colormap="black-white",
-                                               plane_orientation='%s_axes' % axis_name)
-        return ipw
-
-    def make_line_3d(self, axis_name):
-        if axis_name == "x":
-            normal = np.array([1, 0, 0])
-        elif axis_name == "y":
-            normal = np.array([0, 1, 0])
-        elif axis_name == "z":
-            normal = np.array([0, 0, 1])
-        white = (1, 1, 1)
-        grid_size = np.asarray(self.grid.shape)
-        point1 = 0.5 * grid_size * (1 - normal)
-        point2 = grid_size * normal + point1
-        x = [point1[0], point2[0]]
-        y = [point1[1], point2[1]]
-        z = [point1[2], point2[2]]
-        line = mlab.plot3d(x, y, z, color=white, tube_radius=1., figure=self.scene3d.mayavi_scene)
-        return line
-
-    def _ipw_3d_x_default(self):
-        return self.make_ipw_3d('x')
-
-    def _ipw_3d_y_default(self):
-        return self.make_ipw_3d('y')
-
-    def _ipw_3d_z_default(self):
-        return self.make_ipw_3d('z')
-
-    def _line_3d_x_default(self):
-        return self.make_line_3d('x')
-
-    def _line_3d_y_default(self):
-        return self.make_line_3d('y')
-
-    def _line_3d_z_default(self):
-        return self.make_line_3d('z')
+    # def _line_3d_x_default(self):
+    #     return self.make_line_3d('x')
+    #
+    # def _line_3d_y_default(self):
+    #     return self.make_line_3d('y')
+    #
+    # def _line_3d_z_default(self):
+    #     return self.make_line_3d('z')
 
     # def _ipw_pc_default(self):
     #     return self.display_space_cloud()
-    #
+
     # def _ipw_sel_default(self):
     #     return self.display_selections()
-    #
+
     # def _ipw_map_default(self):
     #     return self.display_map()
 
@@ -277,18 +227,13 @@ class VolumeSlicer(HasTraits):
         #                                 figure=self.scene3d.mayavi_scene,
         #                                 )
         self.scene3d.mlab.view(40, 50)
-        # Interaction properties can only be changed after the scene
-        # has been created, and thus the interactor exists
-        for ipw in (self.ipw_3d_x, self.ipw_3d_y, self.ipw_3d_z):
-            # Turn the interaction off
-            ipw.ipw.interaction = 0
         self.scene3d.scene.background = (0, 0, 0)
         # Keep the view always pointing up
         self.scene3d.scene.interactor.interactor_style = \
             tvtk.InteractorStyleTerrain()
 
         # Pick event
-        self.scene3d.scene.interactor.add_observer('KeyPressEvent', self.on_pick)
+        # self.scene3d.scene.interactor.add_observer('KeyPressEvent', self.on_pick)
         self.picker = self.scene3d.mayavi_scene.on_mouse_pick(self.picker_callback)
         # Decrease the tolerance, so that we can more easily select a precise
         # point.
@@ -305,7 +250,7 @@ class VolumeSlicer(HasTraits):
                                                figure=self.scene3d.mayavi_scene)
         scatter = mlab.pipeline.glyph(scatter,
                                       opacity=self.opacity,
-                                      scale_mode='none', scale_factor=3, mode='sphere', colormap="viridis",
+                                      scale_mode='none', scale_factor=10., mode='sphere', colormap="viridis",
                                       figure=self.scene3d.mayavi_scene)
         scatter.actor.actor.pickable = 0
         setattr(self, 'ipw_pc', scatter)
@@ -314,7 +259,7 @@ class VolumeSlicer(HasTraits):
     def display_selections(self):
         data = self.data[list(self.selections.values())]
         scatter = mlab.points3d(data[:, 2], data[:, 1], data[:, 0],
-                                scale_mode='none', scale_factor=3, mode='sphere', color=(1, 1, 1),
+                                scale_mode='none', scale_factor=10., mode='sphere', color=(1, 1, 1),
                                 figure=self.scene3d.mayavi_scene)
         setattr(self, 'ipw_sel', scatter)
 
@@ -324,7 +269,7 @@ class VolumeSlicer(HasTraits):
 
     def addLabel(self, label):
         data = self.data[self.selections[label]]
-        text = mlab.text3d(data[2] + 5., data[1] + 5., data[0] + 5., label, scale=3, color=(1, 1, 1),
+        text = mlab.text3d(data[2] + 0.1, data[1] + 0.1, data[0] + 0.1, label, scale=10., color=(1, 1, 1),
                            figure=self.scene3d.mayavi_scene)
         return text
 
@@ -341,71 +286,6 @@ class VolumeSlicer(HasTraits):
         volume.contour.auto_update_range = True
         setattr(self, 'ipw_map', volume)
 
-    def make_side_view(self, axis_name):
-        scene = getattr(self, 'scene_%s' % axis_name)
-
-        # To avoid copying the data, we take a reference to the
-        # raw VTK dataset, and pass it on to mlab. Mlab will create
-        # a Mayavi source from the VTK without copying it.
-        # We have to specify the figure so that the data gets
-        # added on the figure we are interested in.
-        outline = mlab.pipeline.outline(
-            self.data_src3d.mlab_source.dataset,
-            figure=scene.mayavi_scene,
-        )
-        ipw = mlab.pipeline.image_plane_widget(
-            outline,
-            colormap="viridis",
-            plane_orientation='%s_axes' % axis_name)
-        setattr(self, 'ipw_%s' % axis_name, ipw)
-
-        # Synchronize positions between the corresponding image plane
-        # widgets on different views.
-        ipw.ipw.sync_trait('slice_position',
-                           getattr(self, 'ipw_3d_%s' % axis_name).ipw)
-
-        # Make left-clicking create a crosshair
-        ipw.ipw.left_button_action = 0
-
-        # Add a callback on the image plane widget interaction to
-        # move the others
-        def move_view(obj, evt):
-            # Update 3D plane position
-            position = obj.GetCurrentCursorPosition()
-            self.moveViewToPosition(position, axis_name)
-
-
-
-        ipw.ipw.add_observer('InteractionEvent', move_view)
-        ipw.ipw.add_observer('StartInteractionEvent', move_view)
-
-        # Center the image plane widget
-        ipw.ipw.slice_position = 0.5 * self.grid.shape[
-            self._axis_names[axis_name]]
-
-        # Position the view for the scene
-        views = dict(x=(0, 90),
-                     y=(90, 90),
-                     z=(0, 0),
-                     )
-        scene.mlab.view(*views[axis_name])
-        # 2D interaction: only pan and zoom
-        scene.scene.interactor.interactor_style = \
-            tvtk.InteractorStyleImage()
-        scene.scene.background = (0, 0, 0)
-
-    @on_trait_change('scene_x.activated')
-    def display_scene_x(self):
-        return self.make_side_view('x')
-
-    @on_trait_change('scene_y.activated')
-    def display_scene_y(self):
-        return self.make_side_view('y')
-
-    @on_trait_change('scene_z.activated')
-    def display_scene_z(self):
-        return self.make_side_view('z')
-
     @on_trait_change("opacity")
     def change_opacity_point_cloud(self):
         self.ipw_pc.actor.property.opacity = self.opacity
@@ -419,34 +299,18 @@ class VolumeSlicer(HasTraits):
             contour = (val_max - val_min) * self.threshold + val_min
             volume.contour.contours = [contour]
 
-    @on_trait_change("selection_name")
-    def change_selection_label(self):
-        self.ipw_label.trait_set(text=self.selection_name)
-        self.ipw_label.actor.property.color = (0, 0.76, 1)
-
     @on_trait_change("n_clusters")
     def change_n_clusters(self):
         if not str(self.n_clusters).isdigit():
             self.n_clusters = "10"
 
-    @on_trait_change("show_cut_planes")
-    def update_cut_plane_visibility(self):
-        self.ipw_3d_x.visible = self.show_cut_planes
-        self.ipw_3d_y.visible = self.show_cut_planes
-        self.ipw_3d_z.visible = self.show_cut_planes
-
-    # @on_trait_change("show_selections")
-    # def update_cut_plane_visibility(self):
-    #     self.ipw_sel.visible = self.show_selections
-    #     self.ipw_label.visible = self.show_selections
-
-    def on_pick(self, obj, evt):
-        if obj.GetKeyCode() == "c":
-            pos = np.asarray([self.cursor_position[2],
-                              self.cursor_position[1],
-                              self.cursor_position[0]]).reshape(1, -1)
-            _, ind = self.kdtree_data.query(pos, k=1)
-            self.selections.append(self.z_space[ind, :])
+    # def on_pick(self, obj, evt):
+    #     if obj.GetKeyCode() == "c":
+    #         pos = np.asarray([self.cursor_position[2],
+    #                           self.cursor_position[1],
+    #                           self.cursor_position[0]]).reshape(1, -1)
+    #         _, ind = self.kdtree_data.query(pos, k=1)
+    #         self.selections.append(self.z_space[ind, :])
 
     def _save_selections_fired(self):
         pathFile = os.path.join(self.path, self.save_file)
@@ -456,17 +320,7 @@ class VolumeSlicer(HasTraits):
                     vector = self.z_space[idc]
                     fid.write(' '.join(map(str, vector.reshape(-1))) + "\n")
         self.saveSelectionsDict()
-
-    def _add_selection_fired(self):
-        pos = np.asarray([self.cursor_position[2],
-                          self.cursor_position[1],
-                          self.cursor_position[0]]).reshape(1, -1)
-        _, ind = self.kdtree_data.query(pos, k=1)
-        key = "conf_%d" % (len(self.selections) + 1)
-        self.selections[key] = ind[0, 0]
-
-        data = self.data[list(self.selections.values())]
-        self.ipw_sel.mlab_source.reset(x=data[:, 2], y=data[:, 1], z=data[:, 0])
+        np.savetxt(os.path.join(self.path, 'kmean_labels.txt'), self.interp_val)
 
     def _compute_kmeans_fired(self):
         # First delete all automatic selections associated to a previous KMeans
@@ -477,28 +331,13 @@ class VolumeSlicer(HasTraits):
         if int(self.n_clusters) > 0:
             clusters = KMeans(n_clusters=int(self.n_clusters), n_init=1).fit(self.z_space)
             centers = clusters.cluster_centers_
-            _, inds = self.kdtree_z_space.query(centers, k=1)
+            self.interp_val = clusters.labels_
+            _, inds = self.kdtree_z_pace.query(centers, k=1)
             for idx, ind in enumerate(inds):
                 self.selections["kmean_%d" % (idx + 1)] = ind[0]
         data = self.data[list(self.selections.values())]
         self.ipw_sel.mlab_source.reset(x=data[:, 2], y=data[:, 1], z=data[:, 0])
-
-    def _remove_selection_fired(self):
-        del self.selections[self.current_sel]
-        self.current_sel = ''
-        data = self.data[list(self.selections.values())]
-        self.ipw_sel.mlab_source.reset(x=data[:, 2], y=data[:, 1], z=data[:, 0])
-        self.ipw_label.trait_set(text=self.current_sel)
-        self.ipw_label.actor.property.color = (1, 1, 1)
-
-    def _rename_selection_fired(self):
-        if self.selection_name not in list(self.selections.keys()):
-            self.selections[self.selection_name] = self.selections.pop(self.current_sel)
-            self.current_sel = self.selection_name
-            self.ipw_label.trait_set(text=self.current_sel)
-            self.ipw_label.actor.property.color = (1, 1, 1)
-            data = self.data[list(self.selections.values())]
-            self.ipw_sel.mlab_source.reset(x=data[:, 2], y=data[:, 1], z=data[:, 0])
+        self.ipw_pc.mlab_source.scalars = self.interp_val
 
     def _scale_pts_fired(self):
         scale_factor = float(self.scale_factor)
@@ -540,30 +379,15 @@ class VolumeSlicer(HasTraits):
                              self.ipw_sel.mlab_source.y[point_id], \
                              self.ipw_sel.mlab_source.z[point_id]
 
-                self.moveViewToPosition([px, py, pz], None)
+                self.moveViewToPosition([px, py, pz])
 
                 # Update label
-                self.ipw_label.trait_set(position=[px + 5., py + 5., pz + 5.], text=self.current_sel)
+                self.ipw_label.trait_set(position=[px + 0.1, py + 0.1, pz + 0.1], text=self.current_sel)
                 self.ipw_label.actor.property.color = (1, 1, 1)
 
     # ---------------------------------------------------------------------------
     # Conversion functions
     # ---------------------------------------------------------------------------
-    def interpolateGrid(self):
-        mesh = pv.PolyData(self.data)
-        mesh["interp_val"] = self.interp_val
-        mesh.set_active_scalars("interp_val")
-        grid = pv.create_grid(mesh, dimensions=(300, 300, 300))
-        self.origin = grid.origin
-        self.sr = grid.spacing
-        interpolated = grid.interpolate(mesh, radius=0.1, sharpness=1)
-        self.grid = interpolated.active_scalars.reshape((300, 300, 300))
-
-    def moveDataToGrid(self):
-        self.data[:, 0] = (self.data[:, 0] - self.origin[0]) / self.sr[0]
-        self.data[:, 1] = (self.data[:, 1] - self.origin[1]) / self.sr[1]
-        self.data[:, 2] = (self.data[:, 2] - self.origin[2]) / self.sr[2]
-
     def getDataAtThreshold(self, data, level):
         return np.asarray(np.where(data >= level))
 
@@ -571,10 +395,7 @@ class VolumeSlicer(HasTraits):
     # Read functions
     # ---------------------------------------------------------------------------
     def readMap(self, file):
-        if getExt(file) == ".mrc":
-            map = ImageHandler().read(file + ":mrc").getData()
-        else:
-            map = ImageHandler().read(file).getData()
+        map = ImageHandler().read(file).getData()
         return map
 
     def readSelectionsDict(self):
@@ -599,57 +420,49 @@ class VolumeSlicer(HasTraits):
     # ---------------------------------------------------------------------------
     # View functions
     # ---------------------------------------------------------------------------
-    def moveViewToPosition(self, position, axis_name):
-        setattr(self, 'cursor_position', np.asarray(position))
-        for other_axis, axis_number in self._axis_names.items():
-            if other_axis == axis_name:
-                continue
-            ipw3d = getattr(self, 'ipw_3d_%s' % other_axis)
-            ipw3d.ipw.slice_position = position[axis_number]
-
+    def moveViewToPosition(self, position):
         # Update 3D line position
-        for axis, normal in zip(["x", "y", "z"], np.eye(3)):
-            line3d = getattr(self, 'line_3d_%s' % axis)
-            grid_size = np.asarray(self.grid.shape)
-            position_np = np.asarray(position)
-            point1 = position_np * (1 - normal)
-            point2 = grid_size * normal + point1
-            x = [point1[0], point2[0]]
-            y = [point1[1], point2[1]]
-            z = [point1[2], point2[2]]
-            line3d.mlab_source.reset(x=x, y=y, z=z)
+        # for axis, normal in zip(["x", "y", "z"], np.eye(3)):
+        #     line3d = getattr(self, 'line_3d_%s' % axis)
+        #     grid_size = np.asarray(self.grid.shape)
+        #     position_np = np.asarray(position)
+        #     point1 = position_np * (1 - normal)
+        #     point2 = grid_size * normal + point1
+        #     x = [point1[0], point2[0]]
+        #     y = [point1[1], point2[1]]
+        #     z = [point1[2], point2[2]]
+        #     line3d.mlab_source.reset(x=x, y=y, z=z)
 
         # Update real time conformation
-            pos = np.asarray([position[0], position[1], position[2]]).reshape(1, -1)
-            _, ind = self.kdtree_data.query(pos, k=1)
+        pos = np.asarray([position[2], position[1], position[0]]).reshape(1, -1)
+        _, ind = self.kdtree_data.query(pos, k=1)
 
-            if self.generate_map is not None:
-                if self.mode == "Zernike3D":
-                    self.generate_map("reference.mrc", "mask.mrc", "deformed.mrc",
-                                      self.path, self.z_space[ind[0], :],
-                                      int(self.class_inputs["L1"]), int(self.class_inputs["L2"]), 32)
-                    self.generated_map = self.readMap(os.path.join(self.path, "deformed.mrc"))
-                elif self.mode == "CryoDrgn":
-                    self.generate_map(self.z_space[ind[0], :], self.class_inputs["weights"],
-                                      self.class_inputs["config"], self.path,
-                                      downsample=int(self.class_inputs["boxsize"]),
-                                      apix=float(self.class_inputs["sr"]))
-                    self.generated_map = self.readMap(os.path.join(self.path, "vol_000.mrc"))
-                elif self.mode == "HetSIREN":
-                    self.generate_map(self.class_inputs["weights"], self.z_space[ind[0], :],
-                                      self.path, step=self.class_inputs["step"])
-                    self.generated_map = self.readMap(os.path.join(self.path, "decoded_map_class_1.mrc"))
-                elif self.mode == "NMA":
-                    self.generate_map(self.class_inputs["weights"], self.z_space[ind[0], :],
-                                      self.path, sr=self.class_inputs["sr"])
-                    self.generated_map = self.readMap(os.path.join(self.path, "decoded_map_class_1.mrc"))
+        if self.generate_map is not None:
+            if self.mode == "Zernike3D":
+                self.generate_map("reference.mrc", "mask.mrc", "deformed.mrc",
+                                  self.path, self.z_space[ind[0], :],
+                                  int(self.class_inputs["L1"]), int(self.class_inputs["L2"]), 32)
+                self.generated_map = self.readMap(os.path.join(self.path, "deformed.mrc"))
+            elif self.mode == "CryoDrgn":
+                self.generate_map(self.z_space[ind[0], :], self.class_inputs["weights"],
+                                  self.class_inputs["config"], self.path, downsample=int(self.class_inputs["boxsize"]),
+                                  apix=float(self.class_inputs["sr"]))
+                self.generated_map = self.readMap(os.path.join(self.path, "vol_000.mrc"))
+            elif self.mode == "HetSIREN":
+                self.generate_map(self.class_inputs["weights"], self.z_space[ind[0], :],
+                                  self.path, step=self.class_inputs["step"])
+                self.generated_map = self.readMap(os.path.join(self.path, "decoded_map_class_1.mrc"))
+            elif self.mode == "NMA":
+                self.generate_map(self.class_inputs["weights"], self.z_space[ind[0], :],
+                                  self.path, sr=self.class_inputs["sr"])
+                self.generated_map = self.readMap(os.path.join(self.path, "decoded_map_class_1.mrc"))
 
-        volume = getattr(self, 'ipw_map')
-        val_max = np.amax(volume.mlab_source.m_data.scalar_data)
-        val_min = np.amin(volume.mlab_source.m_data.scalar_data)
-        volume.mlab_source.reset(scalars=self.generated_map)
-        if val_max == val_min:
-            volume.contour.contours = [0.0001]
+            volume = getattr(self, 'ipw_map')
+            val_max = np.amax(volume.mlab_source.m_data.scalar_data)
+            val_min = np.amin(volume.mlab_source.m_data.scalar_data)
+            volume.mlab_source.reset(scalars=self.generated_map)
+            if val_max == val_min:
+                volume.contour.contours = [0.0001]
 
     # ---------------------------------------------------------------------------
     # The layout of the dialog created
@@ -657,30 +470,13 @@ class VolumeSlicer(HasTraits):
     view = View(HSplit(
         HGroup(
             Group(
-                Item('scene_y',
-                     editor=SceneEditor(scene_class=Scene),
-                     height=250, width=300),
-                Item('scene_z',
-                     editor=SceneEditor(scene_class=Scene),
-                     height=250, width=300),
-                show_labels=False,
-            ),
-            Group(
-                Item('scene_x',
-                     editor=SceneEditor(scene_class=Scene),
-                     height=250, width=300),
                 Item('scene3d',
                      editor=SceneEditor(scene_class=MayaviScene),
                      height=250, width=300),
-                show_labels=False,
+                show_labels=False
             ),
             VGroup(
                 Group(
-                    Group(
-                        Item('show_cut_planes'),
-                        # Item('show_selections'),
-                        columns=2
-                    ),
                     Item('opacity',
                          editor=RangeEditor(format='%.02f', low_name='op_min', high_name='op_max', mode='slider')
                          ),
@@ -699,11 +495,7 @@ class VolumeSlicer(HasTraits):
                     ),
                     Group(
                         Item('save_selections'),
-                        Item('add_selection'),
-                        Item('remove_selection'),
-                        Item('selection_name', show_label=False, editor=TextEditor()),
-                        Item('rename_selection'),
-                        show_labels=False, columns=3
+                        show_labels=False, columns=1
                     ),
                     Group(
                         Item('morphing_choice',
@@ -712,7 +504,7 @@ class VolumeSlicer(HasTraits):
                                          "Random walk": "Random walk"}
                              )),
                         Item('morph_chimerax', style='custom', show_label=False),
-                        columns=2)
+                    columns=2)
                 ),
                 Group(
                     Item('scene_c',
@@ -724,7 +516,7 @@ class VolumeSlicer(HasTraits):
         ),
     ),
         resizable=True,
-        title='Volume Slicer',
+        title='Point cloud clustering',
         icon=os.path.join(os.path.dirname(flexutils.__file__), "icon_square.png")
     )
 
@@ -765,5 +557,5 @@ if __name__ == '__main__':
     input_dict['interp_val'] = interp_val
 
     # Initialize volume slicer
-    m = VolumeSlicer(**input_dict)
+    m = PointCloudView(**input_dict)
     m.configure_traits()
