+++ flexutils/viewers/viewers_data.py
@@ -68,9 +68,10 @@ class FlexDataViewer(pwviewer.Viewer):
 
 
 # Register specific sets in pwem dataviewer.
-DataViewer.registerConfig(flexutils.objects.SetOfParticlesFlex,
-                          config={MODE: MODE_MD})
-DataViewer.registerConfig(flexutils.objects.SetOfVolumesFlex,
-                          config={MODE: MODE_MD})
-DataViewer.registerConfig(flexutils.objects.SetOfAtomStructFlex,
-                          config={MODE: MODE_MD})
+if hasattr(DataViewer, "registerConfig"):
+    DataViewer.registerConfig(flexutils.objects.SetOfParticlesFlex,
+                              config={MODE: MODE_MD})
+    DataViewer.registerConfig(flexutils.objects.SetOfVolumesFlex,
+                              config={MODE: MODE_MD})
+    DataViewer.registerConfig(flexutils.objects.SetOfAtomStructFlex,
+                              config={MODE: MODE_MD})
