+++ flexutils/viewers/xmipp/viewer_reduced_space.py
@@ -93,6 +93,6 @@ class XmippReducedSpaceViewer(ProtocolViewer):
         # Run slicer
         args = "--coords %s --deformation %s" \
                % (file_red_space, file_deformation)
-        program = os.path.join(const.VIEWERS, "viewer_point_cloud.py")
+        program = os.path.join(const.VIEWERS, "point_cloud_viewers", "viewer_point_cloud.py")
         program = flexutils.Plugin.getProgram(program)
         runJob(None, program, args)
